Fix build: project now compiles and all checks pass
The project had never built - every CI push since the initial commit failed
with ~50 Kotlin compile errors, including a raw-string parse error in
VisionRuntime that masked the rest as 'Could not load module'. Also fixed:
a deadlocked unit test, NewApi lint errors against minSdk 26, a versionless
androidTest dependency, and the llama/llama_bridge JNI name mismatch.
Verified: assembleDebug builds a 97 MB APK, testDebugUnitTest passes 8/8,
lint reports 0 errors, and androidTest compiles. CI now provisions NDK and
CMake explicitly.

+++ app/src/androidTest/java/com/example/openvoice/VoicePipelineTest.kt
@@ -95,7 +95,7 @@ class VoicePipelineTest {
     @Test
     fun testPipelineEndToEnd() = runBlocking {
         val testAudio = ShortArray(16000) {
-            (Math.sin(2.0 * Math.PI * 440.0 * it / 16000.0) * 8000).toShort()
+            (Math.sin(2.0 * Math.PI * 440.0 * it / 16000.0) * 8000).toInt().toShort()
         }
 
         // VAD test
